fix(frontend): app editor right click menu
